+++ orcid-web/src/main/java/org/orcid/frontend/web/controllers/NotificationController.java
@@ -17,9 +17,11 @@
 package org.orcid.frontend.web.controllers;
 
 import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
 import java.util.List;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.StringUtils;
@@ -28,6 +30,7 @@ import org.orcid.core.manager.EncryptionManager;
 import org.orcid.core.manager.LoadOptions;
 import org.orcid.core.manager.NotificationManager;
 import org.orcid.core.manager.TemplateManager;
+import org.orcid.core.manager.impl.OrcidUrlManager;
 import org.orcid.core.oauth.OrcidProfileUserDetails;
 import org.orcid.jaxb.model.common.Source;
 import org.orcid.jaxb.model.message.OrcidProfile;
@@ -39,8 +42,8 @@ import org.orcid.jaxb.model.notification.permission.NotificationPermission;
 import org.orcid.persistence.dao.NotificationDao;
 import org.orcid.persistence.jpa.entities.ClientDetailsEntity;
 import org.orcid.persistence.jpa.entities.NotificationAddItemsEntity;
-import org.orcid.persistence.jpa.entities.NotificationEntity;
-import org.orcid.persistence.jpa.entities.ProfileEntity;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -68,6 +71,11 @@ public class NotificationController extends BaseController {
 
     @Resource
     private EncryptionManager encryptionManager;
+    
+    @Resource
+    private OrcidUrlManager orcidUrlManager;
+    
+    private static final Logger LOGGER = LoggerFactory.getLogger(NotificationController.class);
 
     @RequestMapping
     public ModelAndView getNotifications() {
@@ -180,6 +188,45 @@ public class NotificationController extends BaseController {
         }
         notificationManager.setActionedDate(notificationOrcid, id);
         return new ModelAndView("redirect:" + redirectUrl);
+    }
+    
+    
+    @RequestMapping(value = { "/preferences" }, method = RequestMethod.GET)
+    public ModelAndView getNotificationPreferencesWindow() {
+    	return new ModelAndView("unsubscribe");
+    }
+    
+    @RequestMapping(value = "/preferences/{encryptedEmail}", method = RequestMethod.GET)
+    public ModelAndView getNotificationPreferencesWindow(HttpServletRequest request, 
+    		@PathVariable("encryptedEmail") String encryptedEmail) throws Exception {
+        ModelAndView result = null;
+        String decryptedEmail = encryptionManager.decryptForExternalUse(new String(Base64.decodeBase64(encryptedEmail), "UTF-8"));
+        OrcidProfile profile = getEffectiveProfile();
+
+        String primaryEmail = profile.getOrcidBio().getContactDetails().retrievePrimaryEmail().getValue();
+
+        if (decryptedEmail.equals(primaryEmail)) {
+        	result = new ModelAndView("unsubscribe");
+        }
+
+        return result;
+    }
+
+    private String createUnsubscribeLink(String email, HttpServletRequest request) {
+        String urlEncodedEmail = null;
+        try {
+            urlEncodedEmail = URLEncoder.encode(email, "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            LOGGER.debug("Unable to url encode email address: {}", email, e);
+        }
+        StringBuilder resendUrl = new StringBuilder();
+        resendUrl.append(orcidUrlManager.getServerStringWithContextPath(request));
+        resendUrl.append("/preferences");
+        if (urlEncodedEmail != null) {
+            resendUrl.append("/");
+            resendUrl.append(urlEncodedEmail);
+        }
+        return resendUrl.toString();
     }
 
     private void addSourceDescription(Notification notification) {
